refactor design in fragment_home.xml

+++ app/src/main/java/com/fighter/stylestorm/data/models/Location.kt
@@ -5,10 +5,10 @@ import com.google.gson.annotations.SerializedName
 
 data class Location(
     @SerializedName("country") val country: String? = null,
-    @SerializedName("lat") val lat: Double? = null,
+    @SerializedName("lat") val lat: Float? = null,
     @SerializedName("localtime") val localtime: String? = null,
     @SerializedName("localtime_epoch") val localtimeEpoch: Int? = null,
-    @SerializedName("lon") val lon: Double? = null,
+    @SerializedName("lon") val lon: Float? = null,
     @SerializedName("name") val name: String? = null,
     @SerializedName("region") val region: String? = null,
     @SerializedName("tz_id") val tzId: String? = null
